[Core] Add function entry/exit logging
Add a `logging` namespace with utilities.

Add `toString` and `concatAsString` helper functions, and use to
stringify arguments and return values.

Log these at `kDebugApi` level (the lowest severity level), taking care
to only construct the strings if that severity level is enabled in the
logger.

Log any exceptions that occur at `kDebug` level and re-`throw`, on the
assumption that normally Katana will decide whether to report or ignore
exceptions, and so adding more logs is only for debug purposes.

Signed-off-by: David Feltell <david.feltell@foundry.com>