[Lint] Add Clang-Tidy config and fix violations
Copy `.clang-tidy` configuration from the OpenAssetIO repository, with
the only change being to remove the list of suppressed headers (which
are not used in this project).

Fix up the code such that there are no Clang-Tidy warnings based on this
configuration.

In particular, lots of messing with `#include`s to remove unused headers
and add headers that were indirectly included, in order to satisfy the
(relatively new) `misc-include-cleaner` check.

Add a note to the README on formatting/linting.

Signed-off-by: David Feltell <david.feltell@foundry.com>