[Tests] Add initial test skeleton
So far KatanaOpenAssetIO has lacked any automated tests. Figuring out
how to add tests has so far been a tricky proposition. The original
implementation of KatanaOpenAssetIO was prototype code that was
cleaned up and published, and lacked any tests. But now its about time
we added some.

So introduce one possible technique for testing the KatanaOpenAssetIO
plugin.

Luckily Geolib exposes a couple of useful C-linkage functions,
specifically for calling from outside. This includes initialisation of
the Geolib host, and scanning for plugins.

This still requires a Katana license, however.

We load KatanaOpenAssetIO via the plugin manager C API and cast the
returned handle to an instance of the base class. This dance seems to
be the cleanest way of getting an instance of KatanaOpenAssetIO.

A cleaner way would be to use the `DefaultAssetManager` utility.
However, that class is missing some functions, notably the publishing
functions (e.g. `createAssetAndPath`).

Direct linking to the plugin doesn't work due it being a MODULE target.
Even if we forcibly link it, we run into symbol visibility issues.
Theses could possibly be resolved, but the solution becomes much more
complex than the proposed solution.

On the OpenAssetIO side, make use of the Basic Asset Library mock/fake
asset manager, used throughout the OpenAssetIO ecosystem for testing
host applications.

Automate setting up dependencies, paths and environment variables as
much as we can through CMake's CTest.

Switch to linking to `Python::Module` in the main library, otherwise
embedding support is pulled in when tests are enabled, adding a
direct dependency on libpython and causing linker errors.

The tests assume a Unix-like OS and will not work in Windows at the
moment.

Signed-off-by: David Feltell <david.feltell@foundry.com>