[Core] Add runAssetPluginCommand "initialize"
Add an `"initialize"` command to the generic `runAssetPluginCommand()`
method. This allows an OpenAssetIO manager to be re-initialised with
different settings.

However, this will only work for string-valued settings. Luckily,
partial updates to settings are supported, as per the OpenAssetIO API
contract. So as long as only string-valued settings are updated, this
works fine.

In particular, this will be useful when re-initialising BAL with a
different library per test.

Signed-off-by: David Feltell <david.feltell@foundry.com>