[Core] Tolerate non-assets in getAssetDisplayName
Katana often calls `getAssetDisplayName` with non-assets, in particular
the Katana project file path, where it is used for the window title.

So rather than throwing an exception and reporting lots of "ERROR" logs,
allow for strings that are not an entity reference and return the asset
ID as the display name in that case.

Signed-off-by: David Feltell <david.feltell@foundry.com>