[Core] Don't duplicate getAssetFields in getAssetAttributes
`getAssetAttributes` is only used once in Katana itself, in the
Importomatic "casting sheet", where it expects a "type" element in the
returned attributes. This case is not yet supported in
KatanaOpenAssetIO.

Otherwise, the `getAssetAttributes` are arbitrary and the docs don't
make it very clear what they should contain vs. `getAssetFields`.

So for now, remove the explicit querying of name and version, so as not
to duplicate metadata that's available from `getAssetFields`.

Instead, `getAssetAttributes` just returns the OpenAssetIO trait
properties. This will be useful when some unit tests are added
imminently.

Signed-off-by: David Feltell <david.feltell@foundry.com>