[Core] Append manager-driven path to preflight asset ID
A typical Katana publish flow when writing a new file is
`createAssetAndPath()` -> `resolveAsset()` -> `postCreateAsset()`. This
corresponds to an OpenAssetIO flow of `preflight()` -> `resolve(...,
kManagerDriven,...)` -> `register()`.

OpenAssetIO differentiates between a `resolve()` for reading existing
data vs. a `resolve()` for requesting new data (i.e. destination file
path). Further, OpenAssetIO expects the destination path to be provided
to `register()`.

On the other hand, Katana assumes that the asset system somehow
"remembers", or can deduce, that `resolveAsset()` should return a
destination path; and that `postCreateAsset()` "remembers", or can
deduce, the destination path that was used.

We therefore need a way to thread through necessary information for
OpenAssetIO via the AssetAPI.

So preempt the publish flow when `createAssetAndPath()` is called to
greedily `resolve()` with `kManagerDriven` and append the result to the
asset ID. Parse and return this as part of the `getAssetFields()` dict.

Doing this gets around two problems
* `resolveAsset` now "knows" that it should return a destination path,
  which manifests as simply returning the path suffix in the asset ID.
* `postCreateAsset` is provided the path as part of the asset fields,
  which can then be used in the publish strategy class to populate the
  OpenAssetIO traits.

Hence provide the asset fields and `FileUrlPathConverter` to the publish
strategy classes, so they can construct a `LocatableContentTrait`.

Add a test that simulates a publish from LookFileMaterialsOut.

Signed-off-by: David Feltell <david.feltell@foundry.com>