[Core] Add patches for Render node preflight/register
Unlike other publishing operations, performing a Disk Render from the
Render node does not automatically bookend the process with
`createAssetAndPath()` and `postCreateAsset()`. Instead, the user is
expected to select "Pre-"/"Post-Render Publish Asset" menu items
manually before/after performing the render.

Unfortunately this process seems to break Katana's own API contract. In
particular, the asset ID returned from `createAssetAndPath()` is
discarded. This means that the wrong destination path may be used, and
the wrong asset ID may be passed to `postCreateAsset()` (of course,
whether this is a problem depends on the asset manager).

So patch the appropriate functionality in Katana's Python code to
retain the asset ID returned from `createAssetAndPath()` as a new
`preflight` parameter on the Render node, which is then recalled when
deducing the destination entity and path. This is done via means of
a Katana "Startup" plugin script.

Add a RenderAssetWidget widget delegate, which Katana uses to display
the output asset IDs in the Render node Parameters tab. Augment with
detection of changes to the new `preflight` parameter, to trigger a
re-sync of the parent widget. Also add a "Reset" button to allow an
in-flight asset ID to be abandoned.

Since the patch and widget are required for "proper" Render node
behaviour, enable their respective CMake options by default.

Signed-off-by: David Feltell <david.feltell@foundry.com>