[Core] Expand getAssetAttributes to all traits + use ',' separator
Use `,` rather than `_` as the separator when constructing the (flat)
attribute key from the nested trait data. It is more conceivable that a
trait ID or property name would genuinely have an `_` in the name than a
`,`, so using `,` hopefully avoids more potential parsing ambiguities.

Include all of an entity's traits in the dictionary returned from
`getAssetAttributes()`, even if they have no `resolve()`able properties.
This means the full trait set can be determined via the AssetAPI,
rather than just those traits with properties. This will be particularly
useful for writing tests.

Add a test for this, including ensuring that the return value can be
encoded properly as a `GroupAttribute`.

Signed-off-by: David Feltell <david.feltell@foundry.com>