[Core] Add publish metadata for LookFileMaterialsOut
When publishing a Look File from a LookFileMaterialsOut node, we're
given an `outputFormat` in the arguments.

It may be useful for the manager to know what the value of this argument
is. E.g. when `resolve()`ing the destination path, the manager must
return either a path to a file or a path to a directory, depending on
the value of `outputFormat` (i.e. "as archive" vs. "as directory").

The closest match to this parameter in the current MediaCreation traits
library is the MIME type on the `LocatableContent` trait.

So set this MIME type to either `inode/directory`, which is
(more-or-less) standardised, or the custom
`application/vnd.foundry.katana.lookfile`, which is not documented
anywhere and just follows the recommended modern pattern for
vendor-specific MIME types. The custom MIME type may of course change
if/when Katana documents one.

Signed-off-by: David Feltell <david.feltell@foundry.com>