[Core] Add publish metadata for Render
Katana provides metadata when the user clicks 'Pre-'/'Post-Render
Publish Asset'. In particular, colour space, name and file extension.
We should provide these to the manager.

Assume the colour space is a valid OCIO colour space and publish
verbatim.

Assume the `outputName` is useful as a display name, and set both
name and qualified name, rather than leaving one blank. It's not
entirely clear which is the better match.

By default, do not imbue the Deep trait (i.e. assume the trait set of
`BitmapImageResourceSpecification`). But if the file extension is one of
the default deep image types, then imbue the Deep trait after the fact.

Convert the file extension to a MIME type, using unofficial or invented
MIME types where none is available, and call out the source in code
comments.

For the other metadata:
* `filePathTemplate` is the `kRead` path at `preflight()` time, so is
  likely incorrect; and is identical to the path extracted from the
  asset ID at `register()` time, so is superfluous.
* `locationSettings` provides the asset ID at `preflight()` time, so
  is superfluous; and is blank at `register()` time.
* `res` provides the resolution alias, but there is no MediaCreation
  trait for that. The OpenAssetIO philosophy is to focus on metadata
  that is not available, or is expensive to determine, from the file
  itself.
* `view` is always blank.

Signed-off-by: David Feltell <david.feltell@foundry.com>