[Core] Add publish metadata for Katana project files
Currently only populate the (invented) MIME type.

Katana distinguishes between two types of publish, `versionUp=True` vs.
`versionUp=False`. The asset manager is involved in both cases. A good
analogy is git tag vs. git revision. There is currently no
straightforward analogy in OpenAssetIO for this.

So publish the same regardless of the value of `versionUp`, for now.

Similarly for the `publish=True`/`False` flag. The meaning of this is
more mysterious, and is ignored for now.

Signed-off-by: David Feltell <david.feltell@foundry.com>