[Core] Support "versionUp" arg
For a few asset types, most notably Katana scene files, a boolean
"versionUp" argument is provided to `createAssetAndPath()`. For example,
"File"->"Save" sets `versionUp=False`, whereas "File"->"Version Up and
Save" sets `versionUp=True`.

The intention is to communicate to the asset manager whether a new
version is desired, or just a revision (or overwrite) of the same
version.

OpenAssetIO has no explicit understanding of this distinction, but we
can hopefully communicate it using the relationship API.

If "versionUp" is not set or is "True", then continue to use the
working reference returned from `preflight()`.

However, if "versionUp" is "False", then attempt to get a new working
reference using a `kWrite` relationship query.

Specifically, get the explicit version ("stableTag") of the target
entity, construct a relationship query filter predicate using that
version, and perform the relationship query with `kWrite` access mode.

Hopefully this unambiguously indicates to the manager that the host
really wants to target a specific version, rather than create a new
version.

If the asset manager doesn't support this workflow, then continue to
use the working reference returned from `preflight()`.

The BAL mock manager, used in the tests, isn't geared up particularly
well for this workflow, but can just about be made to work. So simplify
the tests slightly, ignoring the exact values of entity references
(which could be something like "bal:///cat/revision/v1?v=2"), but
ensuring the traits/paths are as expected.

Bump the required version of BAL to support this workflow.

Signed-off-by: David Feltell <david.feltell@foundry.com>