[Core] Add Katana-specific traits
Add Katana-specific OpenAssetIO traits, and use them when publishing
new entities.

Without Katana-specific traits, many entities are ambiguous, e.g. Katana
projects and look files are both categorised using a
WorkfileSpecification.

Technically, disambiguation can be done using MIME types, since every
published entity is a file with corresponding MIME in its
LocatableContent trait. However, most of the MIME types are
non-standard, and so we end up with one of the very problems that
OpenAssetIO was designed to solve - having to "know" arbitrary strings
and their meaning.

In addition, certain properties that have no corresponding trait in the
community MediaCreation library may be useful to communicate to a
manager.

It's important that these traits are opt-in, such that a generic
OpenAssetIO manager plugin can still work with Katana. As such, we
continue to use MediaCreation specifications as the primary trait set,
but mix in Katana traits to provide an additional optional level of
categorisation.

Signed-off-by: David Feltell <david.feltell@foundry.com>