[Core] Tolerate non-existent entities in getAssetFields
When publishing it is possible that an asset ID points to a non-existent
entity, and the asset manager may flag an `kEntityResolutionError` when
attempting to `resolve()` it. However, `getAssetFields()` will always
be called just prior to publishing, and we should not error in this
case.

So tolerate a non-existent asset by returning the required fields but
with empty values.

Signed-off-by: David Feltell <david.feltell@foundry.com>