[Core] Publish frame range
Katana really deals with image sequences and so when publishing, if we
expect the asset manager to provide a template path to render to, we
should notify it that we're likely rendering a sequence by imbuing the
`FrameRangedTrait`.

Similarly, we should notify the manager of what frames were written by
populating the frame range properties of the `FrameRangedTrait`.
Unfortunately, we cannot in general know at
`preflight()`/`createAssetAndPath()` time which frames will be rendered,
so we must provide that information at `register()`/`postCreateAsset()`
time. Also unfortunately, Katana can't tell us which frames have been
rendered, so we must effectively glob the disk.

Signed-off-by: David Feltell <david.feltell@foundry.com>