[Core] Attempt to disambiguate child projects
When communicating the trait set (type) of an entity to an asset
manager, the `WorkTrait` indicates the entity is a Katana project.
However, we also use `WorkTrait` for other cases, e.g. LiveGroups,
since that's the closest trait available in the public MediaCreation
traits library. This can be confusing for managers.

For example, when publishing, the manager may wish to disambiguate
between the currently open project versus an artifact being published
from that project, even if they are both classed "Work" objects.

So use `SceneResourceSpecification` for LookFiles. The description of
this specification is "Any entity that holds three-dimensional data in
an independent resource", which mostly fits - a look file is essentially
the scene graph of a 3D scene.

For LiveGroups, Macros and FCurves, retain the `WorkTrait` but mix in
the `ConfigTrait` to qualify that the entity is additional
configuration, rather than the root document.

In future, the MediaCreation library will be broadened to include more
traits and specifications, which should hopefully provide more suitable
mappings.

Signed-off-by: David Feltell <david.feltell@foundry.com>