[Core] Better tolerate resolving non-asset
Katana will often try to call `resolveAsset` without first asking if
`isAssetId`, resulting in an exception when we call
`manager.createEntityReference()`, which at best causes log spam.

Katana expects that an asset plugin will return the asset ID unmodified
if the asset ID is not recognised by the plugin.

So add a check to short-circuit and return the asset ID unmodified if
it is not recognised by the manager.